Three forest attacks

+++ Consumer-Game/Assets/Scripts/NPC/NpcAi.cs
@@ -94,8 +94,8 @@ public class NpcAi : MonoBehaviour, HealthInterface
 
     //Heatlh
     
-    [HideInInspector] public float health { get; set; }
-    public float maxHealth  { get; set; }
+    [HideInInspector] public int health { get; set; }
+    public int maxHealth  { get; set; }
     public virtual void InitializeHealth()
     {
         
@@ -103,7 +103,7 @@ public class NpcAi : MonoBehaviour, HealthInterface
         health = maxHealth;
     }
 
-    public virtual void ApplyDamage(float points)
+    public virtual void ApplyDamage(int points)
     {
         health = Mathf.Clamp(health - points, 0, maxHealth);
         if (health <= 0){
